Add files via upload
I've had a really tough time with trying to solve error messages with my GitBash on uploading repositories. Spent the last 2 hours trying to figure out how to upload this. Please let me know if it is correct. I've had it finished but am always having a hard time with my command line.